Move Lambda data classes to own file
