+++ app/src/main/java/com/thebrianwong/protivity/lambda/LambdaDataClasses.kt
@@ -1,0 +1,12 @@
+package com.thebrianwong.protivity.lambda
+
+data class Content (
+    val content: String
+)
+
+data class ResponseObj(
+    val statusCode: Int,
+    val body: Content,
+)
+
+data class BodyDuration(private val duration: Long)
